refactor(spring): share DI metadata attachment
