+++ gitnexus/src/core/ingestion/languages/kotlin/spring-di.ts
@@ -1,92 +1,37 @@
-import { makeScopeId, type ParsedFile, type ScopeId } from 'gitnexus-shared';
-import type { KnowledgeGraph } from '../../../graph/types.js';
-import type { DiInjectionMatch, DiProviderMatch } from '../../di-extractors/index.js';
+import { makeScopeId } from 'gitnexus-shared';
+import { parseSpringInjectionType } from '../../di-extractors/spring.js';
 import {
-  parseSpringInjectionType,
-  SPRING_DI_CAPTURED_FIELD_PROPERTY,
-  SPRING_DI_INJECTION_SITES_PROPERTY,
-  SPRING_DI_PROVIDER_PROPERTY,
-} from '../../di-extractors/spring.js';
-import { createSpringAnnotationNameResolver } from '../../frameworks/spring/bean-candidates.js';
-import { SPRING_BEAN_STEREOTYPES } from '../../frameworks/spring/bean-catalog.js';
-import type { ScopeResolutionIndexes } from '../../model/scope-resolution-indexes.js';
-import { resolveDefGraphId } from '../../scope-resolution/graph-bridge/ids.js';
-import type { GraphNodeLookup } from '../../scope-resolution/graph-bridge/node-lookup.js';
+  createSpringDiMetadataAttacher,
+  hasSpringDiRelevantAnnotation,
+  hasSpringStereotypeSyntax,
+  type SpringDiAnnotationFact,
+  type SpringDiClassFact,
+  type SpringDiDependencyFact,
+  type SpringDiInjectionSiteFact,
+} from '../../frameworks/spring/di-metadata.js';
 import { nodeToCapture, type SyntaxNode } from '../../utils/ast-helpers.js';
 import { getKotlinSpringDiFacts } from './capture-side-channel.js';
 import { isKotlinPackageSiblingVisibilityIncomplete } from './package-siblings.js';
 
-export interface KotlinAnnotationSyntaxFact {
-  readonly name: string;
-  readonly text: string;
+export interface KotlinAnnotationSyntaxFact extends SpringDiAnnotationFact {
   readonly useSiteTarget?: string;
 }
 
-export interface KotlinSpringDependencyFact {
-  readonly name: string;
-  readonly rawType: string;
-  readonly annotations: readonly KotlinAnnotationSyntaxFact[];
-}
+export type KotlinSpringDependencyFact = SpringDiDependencyFact<KotlinAnnotationSyntaxFact>;
 
-export interface KotlinSpringInjectionSiteFact {
-  readonly kind: 'property' | 'constructor' | 'method';
-  readonly memberName: string;
-  readonly implicitConstructor: boolean;
-  readonly annotations: readonly KotlinAnnotationSyntaxFact[];
-  readonly dependencies: readonly KotlinSpringDependencyFact[];
-}
+type KotlinSpringInjectionSiteKind = 'property' | 'constructor' | 'method';
 
-export interface KotlinSpringDiClassFact {
-  readonly classScopeId: ScopeId;
-  readonly classAnnotations: readonly KotlinAnnotationSyntaxFact[];
-  readonly injectionSites: readonly KotlinSpringInjectionSiteFact[];
-}
+export type KotlinSpringInjectionSiteFact = SpringDiInjectionSiteFact<
+  KotlinAnnotationSyntaxFact,
+  KotlinSpringInjectionSiteKind
+>;
 
-const INJECTION_ANNOTATIONS = new Set([
-  'org.springframework.beans.factory.annotation.Autowired',
-  'jakarta.inject.Inject',
-  'javax.inject.Inject',
-]);
-
-const QUALIFIER_ANNOTATIONS = new Set([
-  'org.springframework.beans.factory.annotation.Qualifier',
-  'jakarta.inject.Named',
-  'javax.inject.Named',
-]);
-
-const PRIMARY_ANNOTATIONS = new Set(['org.springframework.context.annotation.Primary']);
-
-const RESOLVABLE_DI_ANNOTATIONS = new Set([
-  ...SPRING_BEAN_STEREOTYPES.keys(),
-  ...INJECTION_ANNOTATIONS,
-  ...QUALIFIER_ANNOTATIONS,
-  ...PRIMARY_ANNOTATIONS,
-]);
-
-const CAPTURE_RELEVANT_ANNOTATIONS = new Set([
-  'Autowired',
-  'Inject',
-  'Qualifier',
-  'Named',
-  'Primary',
-  'Component',
-  'Service',
-  'Repository',
-  'Controller',
-  'RestController',
-  'Configuration',
-]);
-
-const STEREOTYPE_SIMPLE_NAMES = new Set(
-  [...SPRING_BEAN_STEREOTYPES.keys()].map((name) => annotationSimpleName(name)),
-);
+export type KotlinSpringDiClassFact = SpringDiClassFact<
+  KotlinAnnotationSyntaxFact,
+  KotlinSpringInjectionSiteKind
+>;
 
 const KOTLIN_TYPE_NODES = new Set(['user_type', 'nullable_type', 'function_type']);
-
-function annotationSimpleName(name: string): string {
-  const separator = name.lastIndexOf('.');
-  return separator === -1 ? name : name.slice(separator + 1);
-}
 
 function firstDescendantOfType(node: SyntaxNode, type: string): SyntaxNode | undefined {
   const stack = [...node.namedChildren].reverse();
@@ -133,18 +78,6 @@ function annotationFacts(node: SyntaxNode): KotlinAnnotationSyntaxFact[] {
     facts.push(...annotationsFromModifierContainer(child));
   }
   return facts;
-}
-
-function hasRelevantAnnotation(annotations: readonly KotlinAnnotationSyntaxFact[]): boolean {
-  return annotations.some((annotation) =>
-    CAPTURE_RELEVANT_ANNOTATIONS.has(annotationSimpleName(annotation.name)),
-  );
-}
-
-function hasStereotypeSyntax(annotations: readonly KotlinAnnotationSyntaxFact[]): boolean {
-  return annotations.some((annotation) =>
-    STEREOTYPE_SIMPLE_NAMES.has(annotationSimpleName(annotation.name)),
-  );
 }
 
 function directTypeNode(node: SyntaxNode): SyntaxNode | undefined {
@@ -244,9 +177,9 @@ export function captureKotlinSpringDiClassFact(
     const annotations = annotationFacts(primaryConstructor);
     const implicitConstructor =
       constructorCount === 1 &&
-      hasStereotypeSyntax(classAnnotations) &&
-      !hasRelevantAnnotation(annotations);
-    if (implicitConstructor || hasRelevantAnnotation(annotations)) {
+      hasSpringStereotypeSyntax(classAnnotations) &&
+      !hasSpringDiRelevantAnnotation(annotations);
+    if (implicitConstructor || hasSpringDiRelevantAnnotation(annotations)) {
       injectionSites.push({
         kind: 'constructor',
         memberName: '<primary-constructor>',
@@ -261,9 +194,9 @@ export function captureKotlinSpringDiClassFact(
     const annotations = annotationFacts(constructor);
     const implicitConstructor =
       constructorCount === 1 &&
-      hasStereotypeSyntax(classAnnotations) &&
-      !hasRelevantAnnotation(annotations);
-    if (!implicitConstructor && !hasRelevantAnnotation(annotations)) continue;
+      hasSpringStereotypeSyntax(classAnnotations) &&
+      !hasSpringDiRelevantAnnotation(annotations);
+    if (!implicitConstructor && !hasSpringDiRelevantAnnotation(annotations)) continue;
     injectionSites.push({
       kind: 'constructor',
       memberName: '<secondary-constructor>',
@@ -277,7 +210,7 @@ export function captureKotlinSpringDiClassFact(
     for (const member of body.namedChildren) {
       if (member.type === 'property_declaration') {
         const annotations = annotationFacts(member);
-        if (!hasRelevantAnnotation(annotations)) continue;
+        if (!hasSpringDiRelevantAnnotation(annotations)) continue;
         const dependency = propertyDependency(member);
         if (dependency === null) continue;
         injectionSites.push({
@@ -289,7 +222,7 @@ export function captureKotlinSpringDiClassFact(
         });
       } else if (member.type === 'function_declaration') {
         const annotations = annotationFacts(member);
-        if (!hasRelevantAnnotation(annotations)) continue;
+        if (!hasSpringDiRelevantAnnotation(annotations)) continue;
         const name =
           member.namedChildren.find((child) => child.type === 'simple_identifier')?.text.trim() ??
           '<method>';
@@ -304,38 +237,13 @@ export function captureKotlinSpringDiClassFact(
     }
   }
 
-  if (injectionSites.length === 0 && !hasRelevantAnnotation(classAnnotations)) return null;
+  if (injectionSites.length === 0 && !hasSpringDiRelevantAnnotation(classAnnotations)) return null;
   const classCapture = nodeToCapture('@spring-di.class', classNode);
   return {
     classScopeId: makeScopeId({ filePath, range: classCapture.range, kind: 'Class' }),
     classAnnotations,
     injectionSites,
   };
-}
-
-function staticStringArgument(annotationText: string): string | undefined {
-  const args = annotationText.match(/\((.*)\)$/s)?.[1]?.trim();
-  if (args === undefined) return undefined;
-  const value = args.replace(/^value\s*=\s*/, '').trim();
-  const literal = value.match(/^"((?:\\.|[^"\\])*)"$/s);
-  if (literal === null) return undefined;
-  try {
-    return JSON.parse(`"${literal[1]}"`) as string;
-  } catch {
-    return undefined;
-  }
-}
-
-function defaultBeanName(className: string): string {
-  if (className.length === 0) return className;
-  if (
-    className.length > 1 &&
-    className[0] !== className[0].toLowerCase() &&
-    className[1] !== className[1].toLowerCase()
-  ) {
-    return className;
-  }
-  return className[0].toLowerCase() + className.slice(1);
 }
 
 function isApplicableInjectionAnnotation(
@@ -378,151 +286,14 @@ function parseKotlinSpringInjectionType(rawType: string) {
 }
 
 /** Attach resolved, framework-private DI metadata to Kotlin Class nodes. */
-export function attachKotlinSpringDiMetadata(
-  graph: KnowledgeGraph,
-  parsedFiles: readonly ParsedFile[],
-  nodeLookup: GraphNodeLookup,
-  indexes: ScopeResolutionIndexes,
-): void {
-  const resolveAnnotation = createSpringAnnotationNameResolver(indexes);
-
-  for (const parsed of parsedFiles) {
-    const incomplete = isKotlinPackageSiblingVisibilityIncomplete(parsed.filePath);
-    for (const fact of getKotlinSpringDiFacts(parsed.filePath)) {
-      const classScope = indexes.scopeTree.getScope(fact.classScopeId);
-      if (classScope === undefined || classScope.kind !== 'Class') continue;
-      const classDef = classScope.ownedDefs.find((definition) => definition.type === 'Class');
-      if (classDef === undefined) continue;
-      const graphId = resolveDefGraphId(parsed.filePath, classDef, nodeLookup);
-      if (graphId === undefined) continue;
-      const classNode = graph.getNode(graphId);
-      if (classNode === undefined || classNode.label !== 'Class') continue;
-
-      const resolvedAnnotations = new Map<string, string | undefined>();
-      const resolveFact = (
-        annotation: KotlinAnnotationSyntaxFact,
-        enclosingScope: ScopeId | null = classScope.parent,
-      ): string | undefined => {
-        const cacheKey = `${enclosingScope ?? '<root>'}\0${annotation.name}`;
-        if (resolvedAnnotations.has(cacheKey)) return resolvedAnnotations.get(cacheKey);
-        const resolved = resolveAnnotation(
-          annotation.name,
-          parsed,
-          enclosingScope,
-          RESOLVABLE_DI_ANNOTATIONS,
-          incomplete,
-        );
-        resolvedAnnotations.set(cacheKey, resolved);
-        return resolved;
-      };
-
-      const frameworkAnnotations = Array.isArray(classNode.properties.frameworkAnnotations)
-        ? classNode.properties.frameworkAnnotations.filter(
-            (annotation): annotation is string => typeof annotation === 'string',
-          )
-        : [];
-      if (frameworkAnnotations.length > 0) {
-        const names = new Set<string>();
-        let explicitBeanName: string | undefined;
-        let hasDynamicBeanName = false;
-        let primary = false;
-        for (const annotation of fact.classAnnotations) {
-          const resolved = resolveFact(annotation);
-          if (resolved === undefined) continue;
-          if (SPRING_BEAN_STEREOTYPES.has(resolved)) {
-            const argumentText = annotation.text.match(/\((.*)\)$/s)?.[1]?.trim();
-            if (argumentText !== undefined && argumentText.length > 0) {
-              const staticName = staticStringArgument(annotation.text);
-              if (staticName === undefined) hasDynamicBeanName = true;
-              else if (staticName.length > 0) explicitBeanName = staticName;
-            }
-          }
-          if (QUALIFIER_ANNOTATIONS.has(resolved)) {
-            const qualifier = staticStringArgument(annotation.text);
-            if (qualifier !== undefined) names.add(qualifier);
-          }
-          if (PRIMARY_ANNOTATIONS.has(resolved)) primary = true;
-        }
-        if (explicitBeanName !== undefined) names.add(explicitBeanName);
-        else if (!hasDynamicBeanName) names.add(defaultBeanName(classNode.properties.name));
-        const provider: DiProviderMatch = {
-          names: [...names],
-          ...(primary ? { preferenceReason: 'selected @Primary' } : {}),
-        };
-        classNode.properties[SPRING_DI_PROVIDER_PROPERTY] = provider;
-      }
-
-      const matches: DiInjectionMatch[] = [];
-      const semanticallyOwnedPropertyNames = new Set<string>();
-      for (const site of fact.injectionSites) {
-        let injectionAnnotation: KotlinAnnotationSyntaxFact | undefined;
-        for (const annotation of site.annotations) {
-          if (!isApplicableInjectionAnnotation(annotation, site)) continue;
-          const resolved = resolveFact(annotation, classScope.id);
-          if (resolved !== undefined && INJECTION_ANNOTATIONS.has(resolved)) {
-            injectionAnnotation = annotation;
-            break;
-          }
-        }
-        if (injectionAnnotation === undefined) {
-          if (!site.implicitConstructor || frameworkAnnotations.length === 0) continue;
-        } else if (site.kind === 'property') {
-          semanticallyOwnedPropertyNames.add(site.memberName);
-        }
-
-        for (const dependency of site.dependencies) {
-          const parsedType = parseKotlinSpringInjectionType(dependency.rawType);
-          if (parsedType === null) continue;
-          let qualifierAnnotation: KotlinAnnotationSyntaxFact | undefined;
-          for (const annotation of dependency.annotations) {
-            if (!isApplicableQualifierAnnotation(annotation, site)) continue;
-            const resolved = resolveFact(annotation, classScope.id);
-            if (resolved !== undefined && QUALIFIER_ANNOTATIONS.has(resolved)) {
-              qualifierAnnotation = annotation;
-              break;
-            }
-          }
-          const qualifier =
-            qualifierAnnotation === undefined
-              ? undefined
-              : staticStringArgument(qualifierAnnotation.text);
-          if (qualifierAnnotation !== undefined && qualifier === undefined) continue;
-          const trigger =
-            injectionAnnotation === undefined
-              ? 'constructor'
-              : `@${annotationSimpleName(injectionAnnotation.name)} ${site.kind}`;
-          const location =
-            site.kind === 'property'
-              ? site.memberName
-              : `${site.memberName} parameter ${dependency.name}`;
-          matches.push({
-            targetTypeName: parsedType.targetTypeName,
-            cardinality: parsedType.cardinality,
-            ...(qualifier === undefined
-              ? {}
-              : {
-                  namedSelection: {
-                    name: qualifier,
-                    reason: `qualifier "${qualifier}"`,
-                  },
-                }),
-            reason: `Spring DI: ${trigger} ${location}: ${parsedType.displayType}`,
-          });
-        }
-      }
-      if (matches.length > 0) classNode.properties[SPRING_DI_INJECTION_SITES_PROPERTY] = matches;
-
-      for (const propertyName of semanticallyOwnedPropertyNames) {
-        for (const { def } of classScope.bindings.get(propertyName) ?? []) {
-          if (def.ownerId !== classDef.nodeId) continue;
-          const propertyId = resolveDefGraphId(parsed.filePath, def, nodeLookup);
-          if (propertyId === undefined) continue;
-          const property = graph.getNode(propertyId);
-          if (property?.label === 'Property') {
-            property.properties[SPRING_DI_CAPTURED_FIELD_PROPERTY] = true;
-          }
-        }
-      }
-    }
-  }
-}
+export const attachKotlinSpringDiMetadata = createSpringDiMetadataAttacher<
+  KotlinAnnotationSyntaxFact,
+  KotlinSpringInjectionSiteKind
+>({
+  getFacts: getKotlinSpringDiFacts,
+  isPackageVisibilityIncomplete: isKotlinPackageSiblingVisibilityIncomplete,
+  parseInjectionType: parseKotlinSpringInjectionType,
+  capturedMemberKind: 'property',
+  isInjectionAnnotationApplicable: isApplicableInjectionAnnotation,
+  isQualifierAnnotationApplicable: isApplicableQualifierAnnotation,
+});
